minor changes: have the brackets in the files been changed, the type of Duration (double), HourCost (decimal) has been changed, the Id (uint) has been unified, the name field for the model has become mandatory, the class constructor in the tests has been redesigned for primary, the names of the tests have been changed in accordance with the conventions

+++ CarRental.Domain/InternalData/ComponentClasses/CarModelGeneration.cs
@@ -13,7 +13,7 @@ public class CarModelGeneration
     /// <summary>
     /// Unique identifier of the car model generation
     /// </summary>
-    public required int Id { get; set; }
+    public required uint Id { get; set; }
 
     /// <summary>
     /// Calendar year when this generation of the car model was produced
@@ -35,5 +35,5 @@ public class CarModelGeneration
     /// <summary>
     /// Rental cost per hour for vehicles of this model generation
     /// </summary>
-    public required float HourCost { get; set; }
+    public required decimal HourCost { get; set; }
 }
